build-llvm.py: Add the ability to compile LLVM with PGO
Profile guided optimization [1] can help a lot with performance,
especially when the use case for the generated toolchain is very
specific like ours (we only care about compiling the kernel).

This adds a flag that users can enable, which will build a bootstrap
compiler (stage 1), an instrumented compiler (stage 2), a set of kernels
with the instrumented compiler, and then use the generate profiles to
build the final instrumented compiler.

To test the difference with and without PGO, I built a regular LLVM
toolchain:

$ ./build-llvm.py

and a PGO enabled one (using the same revision):

$ ./build-llvm.py --pgo --no-pull

and built the following kernels:

* arm32 allyesconfig
* arm64 allyesconfig
* powerpc powernv_defconfig
* powerpc ppc44x_defconfig
* x86_64 allyesconfig

The compile times for both the compilers and the kernels are below:

build-llvm.py (without PGO): 14m 17s
build-llvm.py (with PGO): 39m 41s

Kernel build (without PGO): 40m 51s
Kernel build (with PGO): 33m 13s

While the total clang compilation time significantly increases, there is
almost a 20% reduction in kernel build time, which can add up a lot in
daily development (plus it is hidden behind an option by default).

[1]: https://llvm.org/docs/HowToBuildWithPGO.html

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>