kernel/build.sh: Add the ability to build allyesconfig builds
This script is a handy benchmarking tool for testing the time
differences in various clang configurations. I've already used
it to see how much speedup we get with PGO and how much slow
down to expect with assertions enabled.

On Linux 5.2, we need just four patches to make everything build
properly (and two of them are upstream for 5.3), see the patch file
for any issues.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>