build-llvm.py: Add an option to perform shallow clones
There are times when one-time builds are desired (e.g. CI runs). Cloning
the LLVM repo with history is wasteful in such cases because the history
will never be used.

Add an option to perform a shallow clone rather than a full clone to
speed up builds when history and updating are not required.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>