tc-build: Check ref exists before checking out when repo is shallow
Our update workflow is:

1. Fetch the origin remote.

2. Checkout the ref requested by the user.

3. See if the ref is a branch.

4. Pull if it is.

Both full and shallow clones will fail at step 2 when the ref does not
exist. However, a ref might exist for a full clone and not for a shallow
clone and the error message might be confusing.

For example, if a user uses '--shallow-clone' with no '--branch'
argument, they will only get access to the default branch (master)
because of the implicit '--single-branch'. If they were to run the
script with '-b release/10.x' after that, the checkout would fail even
though they have technically supplied what would be a valid ref.

We could just live with this failure since the user did shoot themselves
in the foot by not following the advice in the '--shallow-clone' help
text. However, for a better UX, error before attempting a checkout when
.git/shallow is present and git show-branch errors on the ref. This
allows us to tell the user to either:

1. Remove the 'llvm-project' folder and re-run the script with
   '--shallow-clone' + '-b <branch>' so that '--no-single-branch' is
   passed.

2. Run 'git -C llvm-project fetch --unshallow origin' to get all of the
   history.

3. Just do the management themselves and pass in '--no-update'.

If the repo is unshallow, the 'git checkout' error is sufficient enough
to explain what went wrong.

This also adds a small note to the '--branch' help text that explains
'--shallow-clone' does not allow a commit hash to be passed to it
because GitHub does not support '--depth=1' from arbitrary commits.
Checking if a ref is a commit hash programmatically has its flaws so it
is not explicitly checked for.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>