build-llvm.py: Add a note about full LTO
Using full LTO requires a lot of memory. A machine with 32GB of
memory could not complete a build with '--lto=full' before getting
killed (see the link below).

To try and prevent people reporting issues with the script when it is
really their hardware (or lack of), swap the order of the LTO options
to make the thin option seem preferred and add a note about using full
LTO with less than 64GB of memory.

Link: https://www.packet.com/cloud/servers/c3-small/
References: #59
Closes: #85
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>