kernel: Do not add stage 2 folder to PATH unless doing PGO
This can mess with running the script standalone when there is a stage2
folder (either from running the script normally or doing PGO). We only
need to add this folder to PATH when we are building PGO profiles so add
a flag that build-llvm.py can use to make that known.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>