build-llvm.py: Allow user to specify kernel source location for PGO
Some users would like to use their own kernel source to build kernels
for PGO profiles rather than the script doing it for them (e.g. if they
do not have an internet connection).

The kernel/build.sh already has support for doing this, add an option to
build-llvm.py that allows this.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>