kernel/build.sh: Fix patch application with two PGO kernel runs
If a user runs build-llvm.py with '--pgo kernel-{def,allmod}config', the
script will wait for their input because a patch has already been
applied:

patching file arch/powerpc/include/asm/barrier.h
Reversed (or previously applied) patch detected!  Assume -R? [n]

The script should be totally non-interactive. Unfortunately, GNU patch
returns 1 when 'patch -N' detects a previously applied patch so we need
to wrap its output and parse it for this phrase, which is not fatal in
our context. Every other message is treated as fatal and printed in
full.

git apply is another option but it does not appear to work when patching
untracked files, even with '--no-index' and '--unsafe-paths'.

Fixes: b543d42 ("kernel/build.sh: Update patching logic")
Signed-off-by: Nathan Chancellor <nathan@kernel.org>