Eliminate all calls to as_posix()
It is not immediately obvious why as_posix() is being called in most
situations and it is often unnecessary, as subprocess.Popen() and
friends have supported path-like objects without 'shell=True' since
Python 3.6.

Eliminate all these calls, which simplifies everything. However,
path-like objects are not iterable, so explicit conversions to str are
added in places where they need to be in lieu of calling as_posix(),
which appears to be preferred based on my interpretation of the pathlib
documentation.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>