build-binutils.py: Disable internationalization unconditionally
When building on Alpine Linux, linking fails due to undefined
references to libintl_gettext(). We disable internationalization for
certain architectures but not others. Just disable it unconditionally,
which matches Alpine's APKBUILD. The toolchains produced by this script
are generally meant for local consumption, so this will likely have
little impact.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>