tc-build: Stop building the gold plugin
The only reason that we built the gold plugin in the first place was to
support LTO on the Pixel 3, which used ld.gold instead of ld.lld, unlike
Android 4.14+ and mainline's LTO implementation.

The Pixel 3 has been EOL for several months now, so rip out all the gold
plugin enablement. With that, all the binutils functions in utils can be
moved into build-binutils.py directly, as build-llvm.py does not need to
know about binutils anymore.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>