tc_build: kernel: Ensure -Werror gets disabled on PowerPC
-Werror can cause kernel builds to fail, which is a poor experience when
just building the kernel for profiling coverage, so it is disabled with
'KCFLAGS=-Wno-error'. Unfortunately, arch/powerpc has a separate
configuration option to control building just that subdirectory with
-Werror, which overrides the -Wno-error that was provided.

Disable -Werror for powerpc using the special 'disable-werror.config'
target, which merges the configuration to disable -Werror into the
existing configuration, aligning powerpc with the global -Wno-error
we apply for every other architecture.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>