Update DEFAULT_KERNEL_FOR_PGO to 6.15.0
Unfortunately, we pick up three patches.

d897dae472f267dc17c80544c0e6e51d10e52480 is needed to avoid an error
with randstruct enabled (i.e., allmodconfig) with tip of tree LLVM.

The MIPS IAS revert is an undoing of commit 56651128e2fb ("MIPS: Fix
idle VS timer enqueue") and other dependent changes to workaround
breakage with LLVM 20 and older. This may be dropped eventually by just
not building MIPS with LLVM 20 and older when doing PGO.

Finally, the iwlwifi patch is to workaround an error that appears in ARM
allmodconfig in 6.15.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>