tc_build: Enable '--conservative-instrumentation' for BOLT on Apple Silicon
When performing BOLT on Apple Silicon (such as when building in a UTM
virtual machine), running clang hangs due to instrumentation between
exclusive load and store instructions (see the LLVM issue linked in the
comments for more information).

Detect when running on Apple Silicon and enable the conservative
instrumentation option to avoid hitting this issue.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>